federated-learning/requirements.txt
flwr==1.8.0
torch
fastapi
uvicorn
pandas
numpy
scikit-learn
python-jose[cryptography]
pydantic